feat: implement logging configuration with color formatting and bootstrap setup
